+++ lib/opentracing/tracer.rb
@@ -42,10 +42,12 @@ module OpenTracing
     #   References#CHILD_OF reference to the ScopeManager#active.
     # @param finish_on_close [Boolean] whether span should automatically be
     #   finished when Scope#close is called
-    # @yield [Scope] If an optional block is passed to start_active it will
+    # @yield [Scope] If an optional block is passed to start_active_span it will
     #   yield the newly-started Scope. If `finish_on_close` is true then the
     #   Span will be finished automatically after the block is executed.
-    # @return [Scope] The newly-started and activated Scope
+    # @return [Scope, Object] If passed an optional block, start_active_span
+    #   returns the block's return value, otherwise it returns the newly-started
+    #   and activated Scope
     def start_active_span(operation_name,
                           child_of: nil,
                           references: nil,
@@ -54,7 +56,7 @@ module OpenTracing
                           ignore_active_scope: false,
                           finish_on_close: true)
       Scope::NOOP_INSTANCE.tap do |scope|
-        yield scope if block_given?
+        return yield scope if block_given?
       end
     end
 
@@ -74,15 +76,22 @@ module OpenTracing
     # @param tags [Hash] Tags to assign to the Span at start time
     # @param ignore_active_scope [Boolean] whether to create an implicit
     #   References#CHILD_OF reference to the ScopeManager#active.
-    # @return [Span] the newly-started Span instance, which has not been
-    #   automatically registered via the ScopeManager
+    # @yield [Span] If passed an optional block, start_span will yield the
+    #   newly-created span to the block. The span will be finished automatically
+    #   after the block is executed.
+    # @return [Span, Object] If passed an optional block, start_span will return
+    #  the block's return value, otherwise it returns the newly-started Span
+    #  instance, which has not been automatically registered via the
+    #  ScopeManager
     def start_span(operation_name,
                    child_of: nil,
                    references: nil,
                    start_time: Time.now,
                    tags: nil,
                    ignore_active_scope: false)
-      Span::NOOP_INSTANCE
+      Span::NOOP_INSTANCE.tap do |span|
+        return yield span if block_given?
+      end
     end
 
     # Inject a SpanContext into the given carrier